# docs/swagger.yaml
components:
  schemas:
...
    GeneralResponse:
      properties:
        code:
          default: 0
          description: The code of the response, 0 for success, others for failure
          title: Code
          type: integer
        data:
          anyOf:
          - {}
          - type: 'null'
          title: Data
        message:
          description: A message providing more details about the response
          title: Message
          type: string
      required:
      - message
      title: GeneralResponse
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    JWTTokenResponse:
      example:
        jwt_token: jwt_token_string
        username: test@gmail.com
      properties:
        jwt_token:
          anyOf:
          - type: string
          - type: 'null'
          description: The JWT access token for the user
          title: Jwt Token
        username:
          description: The email of the user as username
          title: Username
          type: string
      required:
      - username
      - jwt_token
      title: JWTTokenResponse
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  description: All of backend endpoint for Fastapi Backend.
  title: Fastapi Backend API
  version: '1.0'
openapi: 3.0.3
paths:
...
  /api/v1/auth/jwt/token:
    get:
      operationId: jwt_token_api_v1_auth_jwt_token_get
      parameters:
      - in: header
        name: user
        required: false
        schema:
          title: User
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JWTTokenResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Jwt Token
      tags:
      - Auth
...
  /api/v1/auth/logout:
    get:
      operationId: logout_api_v1_auth_logout_get
      parameters:
      - in: header
        name: user
        required: false
        schema:
          title: User
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Logout
      tags:
      - Auth
...
  /api/v1/auth/redirect:
    get:
      operationId: redirect_api_v1_auth_redirect_get
      parameters:
      - in: query
        name: code
        required: true
        schema:
          title: Code
          type: string
      - in: query
        name: state
        required: true
        schema:
          title: State
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Redirect
      tags:
      - Auth
...